pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "recommender-system"
version = "0.1.0"
description = "Movie recommender system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy<2",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "streamlit>=1.52.1",
    "scikit-surprise>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "black>=24.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
